refactor: 내부 인증 객체 userId 값 변경(Long -> Integer), Getter 추가

+++ gateway/src/main/java/com/two_ddang/logistics/gateway/infrastructure/JwtAuthenticationFilter.java
@@ -82,7 +82,7 @@ public class JwtAuthenticationFilter implements WebFilter {
     }
 
     private PassPort jwtToPassport(Claims claims) {
-        Long userId = claims.get("userId", Long.class);
+        Integer userId = claims.get("userId", Integer.class);
         String username = claims.get("username", String.class);
         String userType = claims.get("userType", String.class);
